requirements.txt
asgiref==3.8.1
Django==5.1.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.2.2
pillow==10.4.0
PyJWT==2.9.0
setuptools==75.1.0
sqlparse==0.5.1
